RAB-1301: Split AttributeGroupRow and AttributeGroupList
