+++ src/Akeneo/Pim/Structure/Bundle/Resources/workspaces/settings-ui/src/components/datagrids/attribute-groups/index.ts
@@ -1,2 +1,2 @@
-export * from './AttributeGroupsDataGrid';
+export * from './AttributeGroupList';
 export * from './MassDeleteAttributeGroupsModal';
